+++ examples/hello_world.js
@@ -1,7 +1,7 @@
 /**
  * Run the following example to perform a recursive crawl of a website using Puppeteer.
  *
- * To run this example on the Apify Platform, select the `Node.js 8 + Chrome on Debian (apify/actor-node-chrome)` base image
+ * To run this example on the Apify Platform, select the `Node.js 10 + Chrome on Debian (apify/actor-node-chrome)` base image
  * on the source tab of your actor configuration.
  */