+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+NEXT
+====================
+- Add `SessionPool` implemenation to `CheerioCrawler`.
+- Fix `Request` constructor not making a copy of objects such as `userData` and `headers`.
+- Fix `desc` option not being applied in local `dataset.getData()`.
+
 0.17.0 / 2019-11-25
 ====================
 - **BREAKING**: Node 8 and 9 are no longer supported. Please use Node 10.17.0 or higher.